opg_project/requirements.txt
asgiref==3.2.3
decorator==4.4.1
Django==3.0
django-crispy-forms==1.8.1
Faker==3.0.0
Jinja2==2.10.3
MarkupSafe==1.1.1
networkx==2.4
psycopg2==2.8.4
Py-YAML-Fixtures==0.5.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
